Changed the timestamp variable to a double so that it can be converted into a timeinterval

+++ League Companion App/Model/mmrTypes.swift
@@ -14,7 +14,7 @@ struct MMR: Codable {
 
 struct Ranked: Codable {
     var avg: Int?
-    var timestamp: Int?
+    var timestamp: Double?
 }
  
 struct Normal: Codable {
